+++ R/AllClassUnions.R
@@ -1,3 +1,3 @@
 #' @importClassesFrom S4Vectors DataFrame
-setClassUnion("data.frameOrSimilar", c("data.frame", "matrix", "DataFrame"))
+setClassUnion("data.frameOrSimilar", c("data.frame", "matrix", "DataFrame", "SummarizedExperiment"))
 setClassUnion("adductClass", c("character", "data.frame"))